Calculate possible joins, and plan for executing joins.
This is pretty simply done at the minute; candidate joins are made by permuting the sources available then checking if it works. At least 1. it's a lazy sequence, so selecting the first one is not expensive; 2. the simplest joins come first.
The planning and execution are pretty simple too, just relational algebra on sequences. Obviously fetches could be done in parallel and doubtless other such things.